Update Auth
Added utils to runtime alias under `#auth/utils` so you may utilize it when using your own middleware.

Moved type back to types. (module build is still weird with using the actual `types.d.ts` file so I'm just gonna leave this alone)

Removed `auth.plugin.mjs` and made it in string format for `addTemplate`. This is more aesthetically pleasing for me as my editor doesn't mingle well with the templating format.

Global middleware does not work, but the option is still there in the off chance Nuxt updates and allows plugins to be accessible when global middleware is enabled. Also added the option to enable or disable the middleware entirely so you may make your own middleware.